+++ internal/controller/pipelinerollout_controller.go
@@ -851,3 +851,29 @@ func (r *PipelineRolloutReconciler) ErrorHandler(pipelineRollout *apiv1.Pipeline
 	r.customMetrics.PipelinesSyncFailed.WithLabelValues().Inc()
 	r.recorder.Eventf(pipelineRollout, corev1.EventTypeWarning, reason, msg+" %v", err.Error())
 }
+
+func parsePipelineStatus(obj *kubernetes.GenericObject) (numaflowv1.PipelineStatus, error) {
+	if obj == nil || len(obj.Status.Raw) == 0 {
+		return numaflowv1.PipelineStatus{}, nil
+	}
+
+	var status numaflowv1.PipelineStatus
+	err := json.Unmarshal(obj.Status.Raw, &status)
+	if err != nil {
+		return numaflowv1.PipelineStatus{}, err
+	}
+
+	return status, nil
+}
+
+func isPipelineReady(status numaflowv1.Status) bool {
+	if len(status.Conditions) == 0 {
+		return false
+	}
+	for _, c := range status.Conditions {
+		if c.Status != metav1.ConditionTrue {
+			return false
+		}
+	}
+	return true
+}
